battle,party: EV-aware CalculateMonStats for a levelled-up retained mon (I-6, #384) (#408)
* battle,party: EV-aware CalculateMonStats for a levelled-up retained mon (I-6, #384)

party::merge_into_save_pokemon filed a levelled-up lead's stat block from
BattlePokemon::stats -- always 0 EVs, per battle's own module docs -- even
when the retained record it replaced carried real EVs in PokemonSubstruct2,
so a loaded EV-trained lead that levelled up this session was filed weaker
than its own EV bytes justify.

battle::calc_stat/calc_max_hp gain the ev/4 term CALC_STAT's arithmetic
always had a documented placeholder for, folded inside the parenthesised
sum exactly where the doc comment prescribed. compute_stats_with_evs
exposes the full formula for a caller with real EVs and no BattlePokemon
of its own to attach them to; compute_stats (every BattlePokemon this
crate builds) stays 0-EV, delegating to it with Evs::default().

party::merge_into_save_pokemon's recompute branch now reads the retained
record's own EV bytes back out of PokemonSubstruct2 and feeds them to
compute_stats_with_evs, rather than the battler's 0-EV cache -- the
retained-block branch and its guard are untouched.

The recompute branch's current-HP write stayed the plain, untranslated
overlay_current_hp, which is capped at the model's own 0-EV maximum: a
full-health lead that levelled up this session filed weaker than the
maximum it was just given, the same corruption shape issue #344 exists to
stop, just reached through the new branch instead of the old unconditional
overwrite. It now re-overlays overlay_current_hp_over_retained_block
against the freshly recomputed maximum and keeps hp_hidden_by_load rather
than retiring it, so the very next save -- once this record's own species
and level land it on the retained branch -- rebases from the same hidden
points instead of forgetting them.

Round 2: keeping the offset unchanged across a recompute was still wrong
whenever the gap between the EV-aware maximum and the 0-EV floor is not
the same size at the old level as at the new one -- CALC_STAT's ev/4 term
is scaled by * level / 100, so it usually is not (a Treecko stored 41/41
at level 13 filed 43/44 at level 14). The recompute branch now rebases
hp_hidden_by_load by the difference between that gap at base.level (the
record it is overwriting, via the new zero_ev_max_hp helper) and at
mon.level() (the block it just recomputed), instead of carrying it
forward unchanged. A fainted-guard test pins the recompute branch's
current-HP write never resurrecting a fainted lead, and the
SENTINEL_EVS_AND_CONDITION fixture's total EVs come down from 790 (a
shape no upstream save can hold) to exactly the 510 cap while keeping the
sp_attack/sp_defense split wide enough to still catch a field swap.

evs_from_substruct2 gets a direct unit test pinning each byte to its named
field, and the sp_attack/sp_defense sentinel bytes shared fixtures already
used are spread wide enough that CALC_STAT's `/ 4 * level / 100` truncation
can no longer make a swap of those two fields pass unnoticed. calc_max_hp's
doc comment cited the wrong pokemon.c lines for its two formula halves;
corrected against the upstream source. Stale test/module doc language left
over from before this issue -- claims that this port cannot rebuild any EV
contribution, and that a recompute retires or merely carries forward the
load-clamp offset -- is brought in line with what the code now does.

Round 3: party::hp_hidden_by_load and white_out's own re-measurement of it
both floored against whatever level the battler happened to be at when
they ran -- the battler's current (possibly already level-up'd or
reconciled-forward) level, not the record's own stored level byte that
merge_into_save_pokemon's recompute branch rebases against as base.level.
A level-up between load and a white-out re-measured the offset against
lead.stats().max_hp (the 0-EV floor at the new level) while the retained
max_hp field it was compared against was still the old level's, so the
very next save lost the gap F(mon.level()) - F(base.level) and filed a
freshly healed lead as damaged -- up to ~63 HP at high level, the #344
corruption shape again. A record whose level byte its own experience word
contradicts hit the identical mismatch at load, before any level-up: the
offset was measured at the already-reconciled level while gap_old still
used the stored byte.

hp_hidden_by_load now takes a dex and floors at stored.level via
zero_ev_max_hp, matching what the recompute branch already assumes;
white_out's re-measurement now just calls hp_hidden_by_load against the
just-healed record, rather than duplicating the (wrong) computation
inline. Both call sites' existing tests are unaffected, since every one
of them loads with the record's level and the battler's already agreeing.
Two new tests -- a white-out after a level-up, and a load with a
contradictory level byte -- each pin the record filing at full rather
than damaged, and fail without this fix.


* party: rebase the load-clamp offset in both directions (I-6, #384)

`CALC_STAT` truncates the EV-aware product and the `0`-EV one
independently, so the gap between them is not monotonic in level: it
shrinks wherever `ev / 4` bought a point at the old level and buys none
at the new one. At such a level-up `battle` widens the live `current_hp`
by the `0`-EV delta, one point more than upstream's `CalculateMonStats`
would, and a `saturating_sub` that could not carry a negative correction
filed that point -- a Treecko stored at 1 HP came back 3 where upstream
files 2. Carry the offset as `i32`, and floor the translated value at 1
so the new subtraction can never report a live lead as fainted.


---------

+++ crates/pokeemerald-rs/src/flow/overworld_phase.rs
@@ -238,7 +238,12 @@ pub(crate) struct OverworldPhase {
     /// lead is decoded from the save, added back by the merge on every
     /// save, and rewritten by the white-out when it completes a heal on
     /// the record directly. Zero when no lead was loaded from a record.
-    pub(super) lead_hp_hidden_by_load: u16,
+    ///
+    /// Signed because the merge rebases it across a level-up and the EV
+    /// gap it rebases by can shrink, leaving the model's live HP a point
+    /// above what upstream's own EV-aware block would hold
+    /// (`party::merge_into_save_pokemon`, issue #384's round-4 review).
+    pub(super) lead_hp_hidden_by_load: i32,
     /// Whether [`Self::party_lead`] is `None` because `save1.player_party[0]`
     /// would not decode, rather than because the slot is genuinely empty
     /// (issue #353).
